requirements.txt
aiogram>=3.1.1
pytoniq-core>=0.1.10
pytonconnect>=0.3.0
python-dotenv>=1.0.0
qrcode~=7.4.2
redis~=5.0.1

pytonapi~=0.3.6